Models

Value-at-Risk models currently available:

Historical Simulation

ValueAtRisk.HistoricalSimulationVaR — Type

HistoricalSimulationVaR{T} <: VaRModel{T}

A naive Historical Simulation approach in which the VaR estimates is the corresponding quantile of the empirical distribution of returns

EWMA Scaled/Filtered Historical Simulation

ValueAtRisk.EWMAHistoricalSimulationVaR — Type

EWMAHistoricalSimulationVaR{T} <: VaRModel{T}

A scaled/filtered Historical Simulation technique in which conditional volatility is calculated using an Exponentially Weighted Moving Average scheme

(GARCH) Filtered Historical simulation*

ValueAtRisk.FilteredHistoricalSimulationVaR — Type

FilteredHistoricalSimulationVaR{T} <: VaRModel{T}

A technique which fits an ARCHModel to the data and forecasts VaR by combining the one-step ahead conditional mean estimate of the model and the quantile of the empirical distributions of the standardized residuals of our data scaled by the one-step ahead conditional volatility estimate

EWMA RiskMetrics approach

ValueAtRisk.EWMARiskMetricsVaR — Type

EWMARiskMetricsVaR{T} <: VaRModel{T}

The RiskMetrics approach to forecasting Value-at-Risk according to which our data is assumed to be normally distributed with mean zero and conditional volatility calculated based on an Exponentially Weighted Moving Average approach

CAViaR (adaptive,symmetric absolute value, asymmetrics slope)

ValueAtRisk.CAViaR_ad — Type

CAViaR_ad{T} <: CAViaR{T}

Engle and Manganelli's Conditionally Autoregressive Value at Risk, adaptive

ValueAtRisk.CAViaR_sym — Type

CAViaR_sym{T} <: CAViaR{T}

Engle and Manganelli's Conditionally Autoregressive Value at Risk, symmetric absolute value

ValueAtRisk.CAViaR_asym — Type

CAViaR_asym{T} <: CAViaR{T}

Engle and Manganelli's Conditionally Autoregressive Value at Risk, asymmetric slope

ARCH models*

ValueAtRisk.ARCHVaR — Type

ARCHVaR{T} <: VaRModel{T}

Estimate VaR using an autoregressive conditional heteroskedasticity model

Extreme Value Theory

ValueAtRisk.ExtremeValueTheoryVaR — Type

ExtremeValueTheoryVaR{T} <: VaRModel{T}

A VaR forecasting technique that makes use of Peaks Over Theshold technique which originates in Extreme Value Theory

Filtered Extreme Value Theory*

ValueAtRisk.FilteredExtremeValueTheoryVaR — Type

FilteredExtremeValueTheoryVaR{T} <: VaRModel{T}

A technique which fits an ARCHModel to the data and forecasts VaR by finding the quantile function of the innovation terms using Extreme Value Theory the standardized residuals. The VaR forecast combines the one-step ahead conditional mean estimate of the model and the result of the quantile function of the innovation terms scaled by the one-step ahead conditional volatility estimate

For models marked with * an ARCH dynamics specification may be supplied: ARCHSpec